feat: scuffed implementation of plugins
