+++ src/assets/modern-icons.css
