+++ src/entrypoints/scroll-period.js
@@ -1,0 +1,110 @@
+export default defineUnlistedScript(() => {
+  const timetable = document.querySelector("[data-timetable-container] div.scrollable");
+
+  if (window.location.pathname === "/" && document.getElementsByClassName("timetable")[0]) {
+    setTimeout(updateScrollbar, 500);
+    setInterval(updateScrollbar, 10000);
+  }
+
+  function updateScrollbar() {
+    const periodIndex = getCurrentPeriod().index;
+    if (periodIndex) {
+      // console.log("scrolling to period", periodIndex);
+      try {
+        const period = document.querySelector(`.timetable thead tr th:nth-child(${periodIndex})`);
+        timetable.scroll({
+          left: period.offsetLeft - 55, // 16 for perfect alignment
+          behavior: "smooth", // or 'auto' to scroll instantly
+        });
+      } catch (error) {
+        console.error("Error updating scrollbar:", error);
+      }
+    }
+  }
+
+  // Period Utils (content scripts do not have support for modules)
+
+  function getListOfPeriods() {
+    const periods = document.querySelectorAll(".timetable thead tr th");
+    return Array.from(periods).map((_, i) => getPeriodData(i));
+  }
+
+  function getCurrentPeriod() {
+    const periodList = getListOfPeriods();
+    const currentTime = new Date().getTime();
+    const currentPeriod = periodList.find((period) => {
+      if (period.header.time) {
+        const { start, end } = period.header.time;
+        // console.log("start time: ", start.getTime(), "end time: ", end.getTime()); // log the start and end times
+        if (start.getTime() <= currentTime && currentTime <= end.getTime()) {
+          return true;
+        }
+      }
+      return false;
+    });
+    // console.log("currentPeriod: ", currentPeriod); // log the currentPeriod
+    return currentPeriod || null;
+  }
+
+  function getPeriodData(index) {
+    if (typeof index !== "number") {
+      console.error("Period number was not provided or is not a number");
+      return null;
+    }
+
+    index += 1;
+    // Example data structure:
+    // {
+    //   header: {
+    //     name: "Period 1"
+    //     time: "8:30am-8-44am"
+    //   },
+    //   data: {
+    //     name: "Programming - Schooltape"
+    //     link: "/home/page/code/schooltape"
+    //     id: "(-ST-CC)"
+    //     room: "Library"
+    //   }
+    // }
+    const header = document.querySelector(`.timetable thead tr th:nth-child(${index})`);
+    const data = document.querySelector(`.timetable tbody tr td:nth-child(${index}) div:nth-child(1) div:nth-child(1)`);
+
+    // console.log(header, data);
+
+    return {
+      header: {
+        name: header?.childNodes[0]?.textContent.trim() || null,
+        time: extractTimes(header?.querySelector("time")?.textContent.trim()) || null,
+      },
+      data: {
+        name: data?.querySelector("a")?.textContent.trim() || null,
+        link: data?.querySelector("a")?.getAttribute("href") || null,
+        id: data?.querySelector("div:nth-child(2)")?.textContent.trim() || null,
+        room: data?.querySelector("div:nth-child(3)")?.textContent.trim() || null,
+      },
+      index: index,
+    };
+  }
+
+  function extractTimes(periodTime) {
+    try {
+      let times = periodTime.split("–"); // en dash
+      let [start, end] = times.map((time) => {
+        let [hour, minute] = time.split(":");
+        let isAM = time.slice(-2) === "am";
+        hour = parseInt(hour);
+        minute = parseInt(minute.substring(0, 2));
+        if (!isAM && hour !== 12) {
+          hour += 12;
+        }
+        let date = new Date();
+        date.setHours(hour, minute, 0, 0);
+        return date;
+      });
+      return { start, end };
+    } catch (error) {
+      console.error(error);
+      return false;
+    }
+  }
+});
