+++ src/entrypoints/start.content.ts
@@ -1,16 +1,23 @@
 export default defineContentScript({
   matches: ["<all_urls>"],
   runAt: "document_start",
-  cssInjectionMode: 'manual',
   async main() {
     let settings = await globalSettings.getValue();
     let themes = await themeSettings.getValue();
+    let plugins = await pluginSettings.getValue();
+
     if (settings.global && settings.urls.includes(window.location.origin)) {
-      logger.info("[start.content.ts] Schooltape is enabled on this site");
+      // inject themes
       if (themes.toggle) {
         logger.info(themes);
         injectCSS("/assets/catppuccin.css");
         injectCatppuccin(themes.flavour, themes.accent);
+      }
+      // inject plugins
+      if (plugins.toggle) {
+        for (let i = 0; i < plugins.enabled.length; i++) {
+          injectPlugin(plugins.enabled[i], "doc-start");
+        }
       }
     }
   }
