+++ src/utils/content-utils.js
@@ -28,36 +28,29 @@ export function injectCatppuccin(flavour, accent) {
     });
 }
 
-// export function injectPlugin(pluginName, loadTime) {
-//   // inject plugins
-//   let xhr = new XMLHttpRequest();
-//   xhr.open("GET", browser.runtime.getURL("/plugins/plugins.json"), true);
-//   xhr.onreadystatechange = function () {
-//     if (xhr.readyState == 4) {
-//       let resp = JSON.parse(xhr.responseText);
-//       // console.log("pluginName", pluginName);
-//       // console.log(resp);
-//       // console.log(resp[pluginName]);
-//       try {
-//         let scripts = resp[pluginName].scripts;
-//         for (let i = 0; i < scripts.length; i++) {
-//           if (scripts[i].execute == loadTime) {
-//             browser.runtime.sendMessage({ inject: scripts[i].path });
-//             // console.log(`Injected Plugin ${i+1}/${scripts.length} (${scripts[i].path})`);
-//           }
-//         }
-//       } catch (error) {
-//         // Enabled plugin not found in plugins.json, remove it from the enabled plugins list
-//         // console.error('An error occurred:', error);
-//         browser.storage.local.get(["settings"], function (data) {
-//           let newSettings = data.settings;
-//           newSettings.enabledPlugins.splice(newSettings.enabledPlugins.indexOf(pluginName), 1);
-//           browser.storage.local.set({ settings: newSettings }, function () {
-//             // console.log(`Removed ${pluginName} from enabled plugins list`);
-//           });
-//         });
-//       }
-//     }
-//   };
-//   xhr.send();
-// }
+export function injectPlugin(pluginName, injectionScript) {
+  // inject plugins
+  let xhr = new XMLHttpRequest();
+  xhr.open("GET", browser.runtime.getURL("plugins.json"), true);
+  xhr.onreadystatechange = async function () {
+    if (xhr.readyState == 4) {
+      let resp = JSON.parse(xhr.responseText);
+      try {
+        let scripts = resp[pluginName].scripts;
+        for (let i = 0; i < scripts.length; i++) {
+          if (scripts[i].execute == injectionScript) {
+            // it's time to inject!
+            browser.runtime.sendMessage({ inject: scripts[i].path });
+          }
+        }
+      } catch (error) {
+        // Enabled plugin not found in plugins.json, remove it from the enabled plugins list
+        logger.warn(`[content-utils] Plugin ${pluginName} not found in plugins.json, disabling`);
+        let plugins = await pluginSettings.getValue();
+        plugins.enabled.splice(plugins.enabled.indexOf(pluginName), 1);
+        await pluginSettings.setValue(plugins);
+      }
+    }
+  };
+  xhr.send();
+}
